Adiciona label modal impressao equipe
