+++ helpers/labelsModalImpressao.js
@@ -1,0 +1,32 @@
+export const labelsModalImpressaoAPS = {
+  titulo: "IMPRESSÃO POR EQUIPES",
+  personalizacaoPrincipal: {
+    titulo: "Deseja imprimir as listas divididas por Equipes?",
+    descricao: "Essa impressão agrupa os cidadãos de acordo com as Equipes correspondentes. Qualquer filtro ou ordenação selecionados anteriormente serão mantidos e aplicados dentro desses grupos.",
+    agrupamentoSim: "Sim, dividir listas por equipes.",
+    agrupamentoNao: "Não, imprimir a lista como ela está.",
+  },
+  personalizacaoSecundaria: {
+    titulo: "Outras opções de impressão por equipes:",
+    recomendacao: "Ideal para distribuir listas para coordenadoras de equipe",
+    separacaoGrupoPorFolha: "Também dividir equipes em folhas separadas",
+    ordenacao: "Também ordenar listas por profissional responsável",
+  },
+  botao: "IMPRIMIR LISTA",
+};
+
+export const labelsModalImpressaoEquipe = {
+  titulo: "IMPRESSÃO POR PROFISSIONAL RESPONSÁVEL",
+  personalizacaoPrincipal: {
+    titulo: "Deseja imprimir as listas divididas por profissional responsável?",
+    descricao: "Essa impressão agrupa os cidadãos de acordo com os profissionais responsáveis correspondentes. Qualquer filtro ou ordenação selecionados anteriormente serão mantidos e aplicados dentro desses grupos.",
+    agrupamentoSim: "Sim, dividir listas por profissional responsável.",
+    agrupamentoNao: "Não, imprimir a lista como ela está.",
+  },
+  personalizacaoSecundaria: {
+    titulo: "Outras opções de impressão por profissional responsável:",
+    recomendacao: "Ideal para distribuir listas para agentes comunitários de saúde",
+    separacaoGrupoPorFolha: "Também dividir a lista impressa com um profissional responsável por folha",
+  },
+  botao: "IMPRIMIR LISTA",
+};
