--- helpers/labelsModalImpressaoAPS.js
@@ -1,16 +1,0 @@
-export const labelsModalImpressaoAPS = {
-  titulo: "IMPRESSÃO POR EQUIPES",
-  personalizacaoPrincipal: {
-    titulo: "Deseja imprimir as listas divididas por Equipes?",
-    descricao: "Essa impressão agrupa os cidadãos de acordo com as Equipes correspondentes. Qualquer filtro ou ordenação selecionados anteriormente serão mantidos e aplicados dentro desses grupos.",
-    agrupamentoSim: "Sim, dividir listas por equipes.",
-    agrupamentoNao: "Não, imprimir a lista como ela está.",
-  },
-  personalizacaoSecundaria: {
-    titulo: "Outras opções de impressão por equipes:",
-    recomendacao: "Ideal para distribuir listas para coordenadoras de equipe",
-    separacaoGrupoPorFolha: "Também dividir equipes em folhas separadas",
-    ordenacao: "Também ordenar listas por profissional responsável",
-  },
-  botao: "IMPRIMIR LISTA",
-};
